Move themed user dataset buttons into own components
