+++ packages/libs/user-datasets/src/lib/Components/Detail/UserDatasetDetail.jsx
@@ -18,10 +18,8 @@ import { isUserDatasetsCompatibleWdkService } from '../../Service/UserDatasetWra
 import SharingModal from '../Sharing/UserDatasetSharingModal';
 import UserDatasetStatus from '../UserDatasetStatus';
 import { makeClassifier, normalizePercentage } from '../UserDatasetUtils';
-import { ThemedGrantAccessButton } from '../List/UserDatasetList';
-
-import { MesaButton, Trash } from '@veupathdb/coreui';
-import { gray, mutedRed } from '@veupathdb/coreui/dist/definitions/colors';
+import { ThemedGrantAccessButton } from '../ThemedGrantAccessButton';
+import { ThemedDeleteButton } from '../ThemedDeleteButton';
 
 import { DateTime } from '../DateTime';
 
@@ -319,50 +317,7 @@ class UserDatasetDetail extends React.Component {
             onPress={this.openSharingModal}
           />
         )}
-        <MesaButton
-          text="Delete"
-          textTransform="none"
-          onPress={this.handleDelete}
-          themeRole={undefined}
-          icon={Trash}
-          styleOverrides={{
-            default: {
-              color: gray[100],
-              textColor: '#4D4D4D',
-              dropShadow: {
-                color: gray[200],
-                blurRadius: '0px',
-                offsetX: '0px',
-                offsetY: '3px',
-              },
-              border: {
-                color: gray[200],
-                style: 'solid',
-                width: 1,
-              },
-            },
-            hover: {
-              color: mutedRed[500],
-              textColor: 'white',
-              dropShadow: {
-                color: mutedRed[700],
-                blurRadius: '0px',
-                offsetX: '0px',
-                offsetY: '4px',
-              },
-            },
-            pressed: {
-              color: mutedRed[600],
-              textColor: 'white',
-              dropShadow: {
-                color: mutedRed[700],
-                blurRadius: '0px',
-                offsetX: '0px',
-                offsetY: '4px',
-              },
-            },
-          }}
-        />
+        <ThemedDeleteButton buttonText="Delete" onPress={this.handleDelete} />
       </div>
     );
   }
