+++ packages/libs/user-datasets/src/lib/Components/List/UserDatasetList.tsx
@@ -37,9 +37,8 @@ import { normalizePercentage, textCell } from '../UserDatasetUtils';
 import './UserDatasetList.scss';
 import { DateTime } from '../DateTime';
 
-import { useUITheme } from '@veupathdb/coreui/dist/components/theming';
-import { MesaButton, Share, Trash } from '@veupathdb/coreui';
-import { gray, mutedRed } from '@veupathdb/coreui/dist/definitions/colors';
+import { ThemedGrantAccessButton } from '../ThemedGrantAccessButton';
+import { ThemedDeleteButton } from '../ThemedDeleteButton';
 
 interface Props {
   baseUrl: string;
@@ -422,50 +421,7 @@ class UserDatasetList extends React.Component<Props, State> {
           userDatasets.forEach((userDataset) => removeUserDataset(userDataset));
         },
         element: (
-          <MesaButton
-            text="Delete"
-            textTransform="none"
-            onPress={() => null}
-            themeRole={undefined}
-            icon={Trash}
-            styleOverrides={{
-              default: {
-                color: gray[100],
-                textColor: '#4D4D4D',
-                dropShadow: {
-                  color: gray[200],
-                  blurRadius: '0px',
-                  offsetX: '0px',
-                  offsetY: '3px',
-                },
-                border: {
-                  color: gray[200],
-                  style: 'solid',
-                  width: 1,
-                },
-              },
-              hover: {
-                color: mutedRed[500],
-                textColor: 'white',
-                dropShadow: {
-                  color: mutedRed[700],
-                  blurRadius: '0px',
-                  offsetX: '0px',
-                  offsetY: '4px',
-                },
-              },
-              pressed: {
-                color: mutedRed[600],
-                textColor: 'white',
-                dropShadow: {
-                  color: mutedRed[700],
-                  blurRadius: '0px',
-                  offsetX: '0px',
-                  offsetY: '4px',
-                },
-              },
-            }}
-          />
+          <ThemedDeleteButton buttonText="Delete" onPress={() => null} />
         ),
         selectionRequired: true,
       },
@@ -692,54 +648,3 @@ class UserDatasetList extends React.Component<Props, State> {
 }
 
 export default UserDatasetList;
-
-type ThemedGrantAccessButtonProps = {
-  buttonText: string;
-  onPress: () => null;
-};
-
-export function ThemedGrantAccessButton({
-  buttonText,
-  onPress,
-}: ThemedGrantAccessButtonProps) {
-  const theme = useUITheme();
-  return (
-    <MesaButton
-      text={buttonText}
-      textTransform="none"
-      onPress={onPress}
-      themeRole="primary"
-      icon={Share}
-      styleOverrides={{
-        default: {
-          border: {
-            color:
-              theme?.palette.primary.hue[theme.palette.primary.level + 100],
-            style: 'solid',
-            width: 1,
-          },
-        },
-        hover: {
-          color: theme?.palette.primary.hue[theme.palette.primary.level + 100],
-          dropShadow: {
-            color:
-              theme?.palette.primary.hue[theme.palette.primary.level + 300],
-            blurRadius: '0px',
-            offsetX: '0px',
-            offsetY: '4px',
-          },
-        },
-        pressed: {
-          color: theme?.palette.primary.hue[theme.palette.primary.level + 200],
-          dropShadow: {
-            color:
-              theme?.palette.primary.hue[theme.palette.primary.level + 400],
-            blurRadius: '0px',
-            offsetX: '0px',
-            offsetY: '4px',
-          },
-        },
-      }}
-    />
-  );
-}
