+++ packages/libs/user-datasets/src/lib/Components/ThemedDeleteButton.tsx
@@ -1,0 +1,54 @@
+import { useUITheme } from '@veupathdb/coreui/dist/components/theming';
+import { MesaButton, Trash } from '@veupathdb/coreui';
+import { gray, mutedRed } from '@veupathdb/coreui/dist/definitions/colors';
+import { ThemedButtonProps } from './ThemedGrantAccessButton';
+
+export function ThemedDeleteButton({ buttonText, onPress }: ThemedButtonProps) {
+  const theme = useUITheme();
+  return (
+    <MesaButton
+      text={buttonText}
+      textTransform="none"
+      onPress={onPress}
+      themeRole={undefined}
+      icon={Trash}
+      styleOverrides={{
+        default: {
+          color: gray[100],
+          textColor: '#4D4D4D',
+          dropShadow: {
+            color: gray[200],
+            blurRadius: '0px',
+            offsetX: '0px',
+            offsetY: '3px',
+          },
+          border: {
+            color: gray[200],
+            style: 'solid',
+            width: 1,
+          },
+        },
+        hover: {
+          color: mutedRed[500],
+          textColor: 'white',
+          dropShadow: {
+            color: mutedRed[700],
+            blurRadius: '0px',
+            offsetX: '0px',
+            offsetY: '4px',
+          },
+        },
+        pressed: {
+          color: mutedRed[600],
+          textColor: 'white',
+          dropShadow: {
+            color: mutedRed[700],
+            blurRadius: '0px',
+            offsetX: '0px',
+            offsetY: '4px',
+          },
+        },
+      }}
+    />
+  );
+}
